# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk mendapatkan daftar stasiun dan jadwal keberangkatan MRT Jakarta.",
        "title": "Jadwal MRT API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1/api",
    "paths": {
        "/stations": {
            "get": {
                "description": "Mengambil seluruh data stasiun MRT Jakarta",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stations"
                ],
                "summary": "Daftar semua stasiun MRT",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/station.GetStationsAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/station.ErrorAPIResponse"
                        }
                    }
                }
            }
        },
        "/stations/{id}": {
            "get": {
                "description": "Mengambil jadwal keberangkatan berikutnya untuk dua jurusan dari stasiun terpilih",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stations"
                ],
                "summary": "Cek jadwal keberangkatan berdasarkan stasiun",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Station ID (nid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/station.GetScedulesAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/station.ErrorAPIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "station.ErrorAPIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "station.GetScedulesAPIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/station.SceduleResponse"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "station.GetStationsAPIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/station.StationResponse"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "station.SceduleResponse": {
            "type": "object",
            "properties": {
                "station": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "station.StationResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}